Fix: Fixed validation for requests

+++ src/controllers/bean.controller.ts
@@ -21,6 +21,9 @@ export class BeanController {
 
   async one(request: Request, response: Response) {
     try {
+      if (typeof request.params.id !== 'string') {
+        return response.status(400).json({ message: 'Invalid bean ID. ID must be a string.' });
+      }
       const bean = await this.beanService.getBeanById(request.params.id);
 
       if (!bean) {
@@ -74,6 +77,9 @@ export class BeanController {
 
   async update(request: Request, response: Response) {
     try {
+      if (typeof request.params.id !== 'string') {
+        return response.status(400).json({ message: 'Invalid bean ID. ID must be a string.' });
+      }
       const {
         name,
         roaster,
@@ -109,6 +115,9 @@ export class BeanController {
 
   async remove(request: Request, response: Response) {
     try {
+      if (typeof request.params.id !== 'string') {
+        return response.status(400).json({ message: 'Invalid bean ID. ID must be a string.' });
+      }
       const success = await this.beanService.deleteBean(request.params.id);
       if (!success) {
         return response.status(404).json({ message: 'Bean not found' });
