+++ src/controllers/user.controller.ts
@@ -21,6 +21,9 @@ export class UserController {
 
   async one(request: Request, response: Response) {
     try {
+      if (typeof request.params.id !== 'string') {
+        return response.status(400).json({ message: 'Invalid user ID. ID must be a string.' });
+      }
       const user = await this.userService.getUserById(request.params.id);
       response.status(200).json(user);
     } catch (error) {
@@ -34,6 +37,9 @@ export class UserController {
 
   async save(request: Request, response: Response) {
     try {
+      if (typeof request.params.id !== 'string') {
+        return response.status(400).json({ message: 'Invalid user ID. ID must be a string.' });
+      }
       const { name, email } = request.body;
 
       const result = await this.userService.createUser({
@@ -52,6 +58,9 @@ export class UserController {
 
   async update(request: Request, response: Response) {
     try {
+      if (typeof request.params.id !== 'string') {
+        return response.status(400).json({ message: 'Invalid user ID. ID must be a string.' });
+      }
       const { name, email } = request.body;
 
       const result = await this.userService.updateUser(request.params.id, {
@@ -73,6 +82,9 @@ export class UserController {
 
   async remove(request: Request, response: Response) {
     try {
+      if (typeof request.params.id !== 'string') {
+        return response.status(400).json({ message: 'Invalid user ID. ID must be a string.' });
+      }
       await this.userService.deleteUser(request.params.id);
       response.status(204).send();
     } catch (error) {
